modified th data loading and implemented trasnforms before training

+++ experiments/unfolding/transforms.py
@@ -300,9 +300,7 @@ class PtPhiEtaE_to_PtPhiEtaM2(BaseTransform):
         zero, one = torch.zeros_like(E), torch.ones_like(E)
         jac_pt = torch.stack((one, zero, zero, -2 * pt * torch.cosh(eta) ** 2), dim=-1)
         jac_phi = torch.stack((zero, one, zero, zero), dim=-1)
-        jac_eta = torch.stack(
-            (zero, zero, one, -(pt**2) * torch.sinh(2 * eta)), dim=-1
-        )
+        jac_eta = torch.stack((zero, zero, one, -(pt**2) * torch.sinh(2 * eta)), dim=-1)
         jac_E = torch.stack((zero, zero, zero, 2 * E), dim=-1)
 
         return torch.stack((jac_pt, jac_phi, jac_eta, jac_E), dim=-1)
@@ -444,7 +442,9 @@ class StandardNormal(BaseTransform):
         self.std = torch.ones(1, 4, device=device, dtype=dtype)
 
     def _forward(self, x):
-        xunit = (x - self.mean.to(x.device)) / self.std.to(x.device)
+        xunit = (x - self.mean.to(x.device, dtype=x.dtype)) / self.std.to(
+            x.device, dtype=x.dtype
+        )
         return xunit
 
     def _inverse(self, xunit):
